Finally got the monkey to move
All my coding was correct (I think) but  the console kept telling me it couldn't GET the monkey image in assets. But I had changed it severy times and it was clearly there in my finder. I then looked on the script in Atom in the assets folder, and it was reading and using a jpg image that I had thought I erased. replacing it with the png directly in the script solved the problem.