Tamed Beyonce
Because I used images, they were not detecting each other because they were still reading as circles. Now they are working as squares and can detect each other. Amen!

Increasing the variable size of each, and within the "if" statements allowed for Beyonce to change size as the came played.