Random Size and Speed change
Arnold avatar is randomly changing size and speed by calculating the position of when Enemy is of the canvas.

I also had to reset the starting position and speed of Arnold avatar for when he went of screen or died. I did this in the "if" statements that describe those parameters.